organization: move completed states to states/ subfolder
