+++ ui/apps/dashboard/src/components/Insights/InsightsDataTable/states/LoadingState.tsx
